fix: move files in correct subfolders
