+++ src/components/buttons/lan/choose-config-btn.js
@@ -1,6 +1,6 @@
 const { ActionRowBuilder, StringSelectMenuOptionBuilder, EmbedBuilder, StringSelectMenuBuilder, MessageFlags } = require("discord.js");
 const { readFileSync } = require("fs");
-const { color } = require("../../../config/config.json");
+const { color } = require("../../../../config/config.json");
 
 module.exports = {
     data: {
