+++ src/components/selectMenus/tournament/select-tournament.js
@@ -1,5 +1,5 @@
 const { ActionRowBuilder, ButtonStyle, ButtonBuilder, EmbedBuilder } = require("discord.js");
-const { color } = require("../../../config/config.json");
+const { color } = require("../../../../config/config.json");
 const { readFileSync } = require("fs");
 
 function generateSlug(str) {
